refactor(recipes): generalize change audit
